Fix FNES broken seeding

+++ FNES-1/fnes1.py
@@ -27,7 +27,7 @@ Make sure to not make any mistakes, though, or your keystreams might come out of
 PS: For security reasons, there are four characters you aren't allowed to encrypt. Sorry!
 """)
 
-tempkey = SHA.new(int(key + time.time() / 10).to_bytes(64, 'big')).digest()[0:16]
+tempkey = SHA.new(int(key + int(time.time() / 10)).to_bytes(64, 'big')).digest()[0:16]
 cipher = ARC4.new(tempkey)
 
 while True:
